Integrate link prediction with main module
